Added Unit Tests & Refactoring
Added unit tests for the refactored MainMenu class which has been put into its own file
