+++ bin/run.rb
@@ -7,7 +7,7 @@ if TTY::Prompt.new().yes?("Would you like to start the program?")
 
     main_menu = MainMenu.new
     main_menu.print_connection_box
-    main_menu.print_prompt
+    main_menu.handle_option(main_menu.print_menu_prompt)
 
 else
     print CLEAR
