+++ config/environment.rb
@@ -4,3 +4,4 @@ Bundler.require
 ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: 'db/development.db')
 require 'net/http'
 require '../lib/run.rb'
+require '../lib/models/MainMenu.rb'
