+++ spec/run_spec.rb
@@ -1,5 +1,4 @@
 require_relative './spec_helper.rb'
-require_relative '../run.rb'
 
 describe 'run_spec' do
